Added some more experimental code

+++ include/Shape.h
@@ -8,6 +8,7 @@
 
 typedef struct coord {
   int x,y,z;
+
 }coord;
 
 typedef struct color {
@@ -29,7 +30,7 @@ class Shape {
   color getColor();
   std::vector<coord> getCoords();
 
-  void rotate(int x, int y, int z);
+  void rotate(double x, double y, double z);
 };
 
 #endif
